update googleplaystoreanalytics folder
