--- gallery/googleplaystore_analytics/references.py
@@ -1,13 +1,0 @@
-import streamlit as st
-
-
-def write():
-    st.title("Acknowledgements")
-    st.write("""
-             Image of Dirichlet Distributions: https://en.wikipedia.org/wiki/Dirichlet_distribution
-             More on Dirichlet Distribution: https://towardsdatascience.com/dirichlet-distribution-a82ab942a879
-             Image of LDA: https://www.depends-on-the-definition.com/understanding-text-data-with-topic-models/
-             Image of DecisionTree: https://forums.tumult.com/t/animating-decision-trees/13870
-             Image of SVM: https://towardsdatascience.com/support-vector-machines-for-classification-fc7c1565e3
-
-             """)
