+++ package/awesome_streamlit/database/resources.py
@@ -5,8 +5,8 @@ Here you maintain the list of awesome resources
 from awesome_streamlit.database import authors, tags
 from awesome_streamlit.database.apps_in_gallery import APPS_IN_GALLERY
 from awesome_streamlit.database.tags import (ALTERNATIVE, APP, APP_IN_GALLERY,
-                                             CODE, SOCIAL, STREAMLIT_TAG,
-                                             TECHNICAL)
+                                             CODE, DEPLOYMENT, SOCIAL,
+                                             STREAMLIT_TAG, TECHNICAL)
 # pylint: disable=line-too-long
 from awesome_streamlit.shared.models import Resource
 
@@ -630,6 +630,12 @@ RESOURCES = (
             name="Uber Data Explorer App",
             url="https://dataexplorerlit.herokuapp.com/",
             tags=[CODE],
+            is_awesome=True,
+        ),
+        Resource(
+            name="Google Playstore Analytics",
+            url="https://playstore-analytics.herokuapp.com/",
+            tags=[CODE, DEPLOYMENT, APP_IN_GALLERY],
             is_awesome=True,
         ),
     ]
